backend/requirements.txt
asgiref==3.8.1
constantly==15.1.0
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.1
et-xmlfile==1.0.1
Faker==26.1.0
google-pasta==0.2.0
hpack==4.0.0
munkres==1.1.4
numpy==1.26.4
protobuf==4.24.4
pyarrow==15.0.0
pycairo==1.26.0
PyGObject==3.46.0
python-dateutil==*
requests-unixsocket==0.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
zstandard==0.22.0
